project/broker/requirements.txt
paho-mqtt
ipdb